requirements.txt
asgiref==3.2.3
Django==3.0.7
django-mdeditor==0.1.18
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-registration-redux==2.7
jsonfield==3.1.0
Markdown==3.3.3
Pillow==7.1.0
pytz==2019.3
sqlparse==0.3.0
swapper==1.1.2.post1
